fixup! Redesign the Dining tab
